xj1corecloud/requirements.txt
paho-mqtt>=1.6.0,<2.0.0
python-dateutil>=2.8.0
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0
